chore: 사용자 정보 불러오기
