+++ src/main/java/com/lucy/arti/oauth/controller/AuthController.java
@@ -1,16 +1,16 @@
 package com.lucy.arti.oauth.controller;
 
 import com.lucy.arti.member.model.MemberRepository;
+import com.lucy.arti.oauth.service.KakaoOauth2;
+import com.lucy.arti.oauth.KakaoUserInfo;
 import com.lucy.arti.oauth.service.AuthService;
 import com.lucy.arti.oauth.dto.TokenDto;
 import com.lucy.arti.oauth.dto.KakaoLoginRequestDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.json.JSONObject;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.HashMap;
-import java.util.Map;
 
 @RestController
 @RequiredArgsConstructor
@@ -30,13 +30,6 @@ public class AuthController {
     public ResponseEntity kakaoLogout(@RequestHeader(name = "Authorization") String bearerToken) {
         System.out.println("bearerToken = " + bearerToken);
         return authService.logout(bearerToken);
-    }
-
-    @GetMapping("")
-    public Map getByAccessToken(@RequestHeader(name = "Authorization") String bearerToken) {
-        Map<String, Long> map = new HashMap<>();
-        map.put("id", memberRepository.findByAccessToken(bearerToken).get().getId());
-        return map;
     }
 }
