+++ src/main/java/com/lucy/arti/oauth/controller/KakaoController.java
@@ -1,7 +1,7 @@
 package com.lucy.arti.oauth.controller;
 
 import com.lucy.arti.oauth.AuthCode;
-import com.lucy.arti.oauth.KakaoOauth2;
+import com.lucy.arti.oauth.service.KakaoOauth2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
